Change primary keys for QuestStep and Player entities from Long to Int

+++ app/src/main/java/ee/ut/cs/tartu_explorer/core/data/local/entities/QuestStepEntity.kt
@@ -32,7 +32,7 @@ import androidx.room.PrimaryKey
 )
 data class QuestStepEntity(
     @PrimaryKey(autoGenerate = true)
-    val id: Long = 0,
-    val questId: Long,
+    val id: Int = 0,
+    val questId: Int,
     val message: String // Dieses Feld hat gefehlt
 )
